backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
boto3==1.38.36
botocore==1.38.36
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
django-storages==1.14.6
django-stubs==5.2.0
django-stubs-ext==5.2.0
djangorestframework==3.16.0
djangorestframework-stubs==3.16.0
dotenv==0.9.9
drf-spectacular==0.28.0
flake8==7.2.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
mypy==1.16.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pycodestyle==2.13.0
pyflakes==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
types-PyYAML==6.0.12.20250516
types-requests==2.32.4.20250611
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
